........S. [ZBX-4507] added possibility to remove a "Discovered hosts" group from a discovered host by an action
git-svn-id: svn://svn.zabbix.com/branches/dev@24864 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-4507/src/zabbix_server/operations.c
@@ -408,34 +408,94 @@ end:
 
 /******************************************************************************
  *                                                                            *
- * Function: add_discovered_host_group                                        *
+ * Function: add_discovered_host_groups                                       *
  *                                                                            *
  * Purpose: add group to host if not added already                            *
  *                                                                            *
  * Author: Alexander Vladishev                                                *
  *                                                                            *
  ******************************************************************************/
-static void	add_discovered_host_group(zbx_uint64_t hostid, zbx_uint64_t groupid)
+static void	add_discovered_host_groups(zbx_uint64_t hostid, zbx_vector_uint64_t *groupids)
 {
+	const char	*__function_name = "add_discovered_host_groups";
+
 	DB_RESULT	result;
 	DB_ROW		row;
-	zbx_uint64_t	hostgroupid;
+	zbx_uint64_t	groupid;
+	char		*sql = NULL;
+	int		sql_alloc = 256, sql_offset = 0, i;
 
-	result = DBselect(
-			"select hostgroupid"
+	zabbix_log(LOG_LEVEL_DEBUG, "In %s()", __function_name);
+
+	sql = zbx_malloc(sql, sql_alloc);
+
+	zbx_snprintf_alloc(&sql, &sql_alloc, &sql_offset, 256,
+			"select groupid"
 			" from hosts_groups"
-			" where groupid=" ZBX_FS_UI64
-				" and hostid=" ZBX_FS_UI64,
-			groupid, hostid);
+			" where hostid=" ZBX_FS_UI64
+				" and",
+			hostid);
+	DBadd_condition_alloc(&sql, &sql_alloc, &sql_offset, "groupid", groupids->values, groupids->values_num);
 
-	if (NULL == (row = DBfetch(result)))
+	result = DBselect("%s", sql);
+
+	while (NULL != (row = DBfetch(result)))
 	{
-		hostgroupid = DBget_maxid("hosts_groups");
-		DBexecute("insert into hosts_groups (hostgroupid,hostid,groupid)"
-				" values (" ZBX_FS_UI64 "," ZBX_FS_UI64 "," ZBX_FS_UI64 ")",
-				hostgroupid, hostid, groupid);
+		ZBX_STR2UINT64(groupid, row[0]);
+
+		if (FAIL == (i = zbx_vector_uint64_bsearch(groupids, groupid, ZBX_DEFAULT_UINT64_COMPARE_FUNC)))
+		{
+			THIS_SHOULD_NEVER_HAPPEN;
+			continue;
+		}
+
+		zbx_vector_uint64_remove_noorder(groupids, i);
 	}
 	DBfree_result(result);
+
+	if (0 != groupids->values_num)
+	{
+		const char	*ins_hosts_groups_sql = "insert into hosts_groups (hostgroupid,hostid,groupid) values ";
+		zbx_uint64_t	hostgroupid;
+#ifdef HAVE_MULTIROW_INSERT
+		const char	*row_dl = ",";
+#else
+		const char	*row_dl = ";\n";
+#endif
+
+		hostgroupid = DBget_maxid_num("hosts_groups", groupids->values_num);
+
+		sql_offset = 0;
+#ifdef HAVE_ORACLE
+		zbx_snprintf_alloc(&sql, &sql_alloc, &sql_offset, 7, "begin\n");
+#endif
+#ifdef HAVE_MULTIROW_INSERT
+		zbx_strcpy_alloc(&sql, &sql_alloc, &sql_offset, ins_hosts_groups_sql);
+#endif
+		for (i = 0; i < groupids->values_num; i++)
+		{
+#ifndef HAVE_MULTIROW_INSERT
+			zbx_strcpy_alloc(&sql, &sql_alloc, &sql_offset, ins_hosts_groups_sql);
+#endif
+			zbx_snprintf_alloc(&sql, &sql_alloc, &sql_offset, 67,
+					"(" ZBX_FS_UI64 "," ZBX_FS_UI64 "," ZBX_FS_UI64 ")%s",
+					hostgroupid++, hostid, groupids->values[i], row_dl);
+
+#ifdef HAVE_ORACLE
+			zbx_snprintf_alloc(&sql, &sql_alloc, &sql_offset, 6, "end;\n");
+#endif
+		}
+
+#ifdef HAVE_MULTIROW_INSERT
+		sql_offset--;
+		zbx_strcpy_alloc(&sql, &sql_alloc, &sql_offset, ";\n");
+#endif
+		DBexecute("%s", sql);
+	}
+
+	zbx_free(sql);
+
+	zabbix_log(LOG_LEVEL_DEBUG, "End of %s()", __function_name);
 }
 
 /******************************************************************************
@@ -453,18 +513,21 @@ static void	add_discovered_host_group(zbx_uint64_t hostid, zbx_uint64_t groupid)
  ******************************************************************************/
 static zbx_uint64_t	add_discovered_host(DB_EVENT *event)
 {
-	const char	*__function_name = "add_discovered_host";
-	DB_RESULT	result;
-	DB_RESULT	result2;
-	DB_ROW		row;
-	DB_ROW		row2;
-	zbx_uint64_t	hostid = 0, proxy_hostid, host_proxy_hostid;
-	char		host[MAX_STRING_LEN], *host_esc, *ip_esc, *host_unique, *host_unique_esc;
-	int		port;
-	zbx_uint64_t	groupid;
+	const char		*__function_name = "add_discovered_host";
 
-	zabbix_log(LOG_LEVEL_DEBUG, "In %s(eventid:" ZBX_FS_UI64 ")",
-			__function_name, event->eventid);
+	DB_RESULT		result;
+	DB_RESULT		result2;
+	DB_ROW			row;
+	DB_ROW			row2;
+	zbx_uint64_t		hostid = 0, proxy_hostid, host_proxy_hostid;
+	char			host[MAX_STRING_LEN], *host_esc, *ip_esc, *host_unique, *host_unique_esc;
+	int			port;
+	zbx_uint64_t		groupid;
+	zbx_vector_uint64_t	groupids;
+
+	zabbix_log(LOG_LEVEL_DEBUG, "In %s() eventid:" ZBX_FS_UI64, __function_name, event->eventid);
+
+	zbx_vector_uint64_create(&groupids);
 
 	result = DBselect(
 			"select discovery_groupid"
@@ -476,11 +539,12 @@ static zbx_uint64_t	add_discovered_host(DB_EVENT *event)
 	if (NULL != (row = DBfetch(result)))
 	{
 		ZBX_STR2UINT64(groupid, row[0]);
+		zbx_vector_uint64_append(&groupids, groupid);
 	}
 	else
 	{
 		zabbix_log(LOG_LEVEL_WARNING, "cannot add discovered host: group for discovered hosts is not defined");
-		return 0;
+		goto clean;
 	}
 	DBfree_result(result);
 
@@ -513,7 +577,7 @@ static zbx_uint64_t	add_discovered_host(DB_EVENT *event)
 					event->objectid);
 			break;
 		default:
-			return 0;
+			goto clean;
 	}
 
 	if (NULL != (row = DBfetch(result)))
@@ -544,6 +608,8 @@ static zbx_uint64_t	add_discovered_host(DB_EVENT *event)
 				DBexecute("insert into hosts (hostid,proxy_hostid,host,useip,dns)"
 						" values (" ZBX_FS_UI64 "," ZBX_FS_UI64 ",'%s',0,'%s')",
 						hostid, proxy_hostid, host_esc, host_esc);
+
+				add_discovered_host_groups(hostid, &groupids);
 			}
 			else
 			{
@@ -607,6 +673,8 @@ static zbx_uint64_t	add_discovered_host(DB_EVENT *event)
 
 				zbx_free(host_unique);
 				zbx_free(host_unique_esc);
+
+				add_discovered_host_groups(hostid, &groupids);
 			}
 			else
 			{
@@ -628,9 +696,8 @@ static zbx_uint64_t	add_discovered_host(DB_EVENT *event)
 		}
 	}
 	DBfree_result(result);
-
-	if (0 != hostid)
-		add_discovered_host_group(hostid, groupid);
+clean:
+	zbx_vector_uint64_destroy(&groupids);
 
 	zabbix_log(LOG_LEVEL_DEBUG, "End of %s()", __function_name);
 
@@ -778,7 +845,7 @@ void	op_host_disable(DB_EVENT *event)
  * Author: Alexei Vladishev                                                   *
  *                                                                            *
  ******************************************************************************/
-void	op_group_add(DB_EVENT *event, zbx_uint64_t groupid)
+void	op_group_add(DB_EVENT *event, zbx_vector_uint64_t *groupids)
 {
 	const char	*__function_name = "op_group_add";
 	zbx_uint64_t	hostid;
@@ -794,7 +861,7 @@ void	op_group_add(DB_EVENT *event, zbx_uint64_t groupid)
 	if (0 == (hostid = add_discovered_host(event)))
 		return;
 
-	add_discovered_host_group(hostid, groupid);
+	add_discovered_host_groups(hostid, groupids);
 
 	zabbix_log(LOG_LEVEL_DEBUG, "End of %s()", __function_name);
 }
@@ -811,10 +878,14 @@ void	op_group_add(DB_EVENT *event, zbx_uint64_t groupid)
  * Author: Alexei Vladishev                                                   *
  *                                                                            *
  ******************************************************************************/
-void	op_group_del(DB_EVENT *event, zbx_uint64_t groupid)
+void	op_group_del(DB_EVENT *event, zbx_vector_uint64_t *groupids)
 {
 	const char	*__function_name = "op_group_del";
+
+	DB_RESULT	result;
 	zbx_uint64_t	hostid;
+	char		*sql = NULL;
+	int		sql_alloc = 256, sql_offset = 0;
 
 	zabbix_log(LOG_LEVEL_DEBUG, "In %s()", __function_name);
 
@@ -827,12 +898,38 @@ void	op_group_del(DB_EVENT *event, zbx_uint64_t groupid)
 	if (0 == (hostid = select_discovered_host(event)))
 		return;
 
-	DBexecute(
-			"delete from hosts_groups"
-			" where hostid=" ZBX_FS_UI64
-				" and groupid=" ZBX_FS_UI64,
-			hostid,
-			groupid);
+	sql = zbx_malloc(sql, sql_alloc);
+
+	zbx_snprintf_alloc(&sql, &sql_alloc, &sql_offset, 256,
+		"select groupid"
+		" from hosts_groups"
+		" where hostid=" ZBX_FS_UI64
+			" and not",
+		hostid);
+	DBadd_condition_alloc(&sql, &sql_alloc, &sql_offset, "groupid", groupids->values, groupids->values_num);
+
+	result = DBselectN(sql, 1);
+
+	if (NULL != DBfetch(result))
+	{
+		sql_offset = 0;
+		zbx_snprintf_alloc(&sql, &sql_alloc, &sql_offset, 256,
+				"delete from hosts_groups"
+				" where hostid=" ZBX_FS_UI64
+					" and",
+				hostid);
+		DBadd_condition_alloc(&sql, &sql_alloc, &sql_offset, "groupid", groupids->values, groupids->values_num);
+
+		DBexecute("%s", sql);
+	}
+	else
+	{
+		zabbix_log(LOG_LEVEL_WARNING, "cannot delete group(s): the \"%s\" host remains without groups",
+				zbx_host_string(hostid));
+	}
+	DBfree_result(result);
+
+	zbx_free(sql);
 
 	zabbix_log(LOG_LEVEL_DEBUG, "End of %s()", __function_name);
 }
